T-47: Merged Develop, fixing services WIP

+++ GauchoGrubAndroid/app/src/main/java/com/g10/gauchogrub/services/DataAutomationService.java
@@ -5,17 +5,16 @@ import android.app.Service;
 import android.content.Intent;
 import android.os.AsyncTask;
 import android.os.IBinder;
-import android.provider.MediaStore;
 
+import com.g10.gauchogrub.menu.DiningCommon;
+import com.g10.gauchogrub.utils.CacheUtils;
 import com.g10.gauchogrub.utils.WebUtils;
 
 import org.joda.time.DateTime;
 
 import java.io.File;
-import java.lang.reflect.Field;
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
-import java.util.Date;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
@@ -59,7 +58,8 @@ public class DataAutomationService extends Service{
             for(File f : cachedFiles) {
                 DateTime dateModified = new DateTime(f.lastModified());
                 DateTime yesterday = DateTime.now().minusDays(1);
-                if(dateModified.isBefore(yesterday)) {
+                //Checks if it is the favorites file
+                if(dateModified.isBefore(yesterday) && !f.getName().contains("Favorites")) {
                     try {
                         f.delete();
                     }
@@ -81,20 +81,25 @@ public class DataAutomationService extends Service{
         @Override
         protected Void doInBackground(Void... params) {
             WebUtils w = new WebUtils();
+            CacheUtils c = new CacheUtils();
             DateTime day = DateTime.now();
             DateFormat requestFormat = new SimpleDateFormat("MM/dd/yyyy");
             DateFormat saveFormat = new SimpleDateFormat("MMddyyyy"); //TODO: Replace with static in CachingUtils class later
-
+            //for each of the 7 days
             for(int i = 0; i < 7; i++) {
-                String requestDate = requestFormat.format(day.plusDays(i));
-                try {
-                    String menu = w.createMenuString("diningCommon", requestDate); //TODO: Replace "diningCommon" once merged with T-47
-                    String saveDate = saveFormat.format(day.plusDays(i));
-                    //TODO: Cache menus
+                //for each dining common
+                for(int j = 0; j < 4; j++) {
+                    String requestDate = requestFormat.format(day.plusDays(i));
+                    try {
+                        String menu = w.createMenuString(DiningCommon.DATA_USE_DINING_COMMONS[j], requestDate);
+                        String saveDate = saveFormat.format(day.plusDays(i));
+                        String fileName = DiningCommon.DATA_USE_DINING_COMMONS[j] + saveDate;
+                        c.cacheFile(getBaseContext(), fileName, menu);
+                    }
+                    catch (Exception ex) {
+                        logger.log(Level.INFO, ex.getMessage());
+                    }
 
-                }
-                catch (Exception ex) {
-                    logger.log(Level.INFO, ex.getMessage());
                 }
             }
             return null;
